move procedure to new design
