+++ app/controllers/admin/procedures_controller.rb
@@ -2,7 +2,7 @@ class Admin::ProceduresController < AdminController
   include SmartListing::Helper::ControllerExtensions
   helper SmartListing::Helper
 
-  before_action :retrieve_procedure, only: [:show, :edit, :delete_logo, :delete_deliberation, :delete_notice, :monavis, :update_monavis, :publish_validate, :publish]
+  before_action :retrieve_procedure, only: [:show, :delete_logo, :delete_deliberation, :delete_notice, :publish_validate, :publish]
 
   def index
     if current_administrateur.procedures.count != 0
@@ -49,9 +49,6 @@ class Admin::ProceduresController < AdminController
     @current_administrateur = current_administrateur
   end
 
-  def edit
-  end
-
   def destroy
     procedure = current_administrateur.procedures.find(params[:id])
 
@@ -64,40 +61,6 @@ class Admin::ProceduresController < AdminController
 
     flash.notice = 'Démarche supprimée'
     redirect_to admin_procedures_draft_path
-  end
-
-  def new
-    @procedure ||= Procedure.new(for_individual: true)
-  end
-
-  def create
-    @procedure = Procedure.new(procedure_params.merge(administrateurs: [current_administrateur]))
-
-    if !@procedure.save
-      flash.now.alert = @procedure.errors.full_messages
-      render 'new'
-    else
-      flash.notice = 'Démarche enregistrée.'
-      current_administrateur.instructeur.assign_to_procedure(@procedure)
-
-      redirect_to champs_procedure_path(@procedure)
-    end
-  end
-
-  def update
-    @procedure = current_administrateur.procedures.find(params[:id])
-
-    if !@procedure.update(procedure_params)
-      flash.now.alert = @procedure.errors.full_messages
-      render 'edit'
-    elsif @procedure.brouillon?
-      reset_procedure
-      flash.notice = 'Démarche modifiée. Tous les dossiers de cette démarche ont été supprimés.'
-      redirect_to edit_admin_procedure_path(id: @procedure.id)
-    else
-      flash.notice = 'Démarche modifiée.'
-      redirect_to edit_admin_procedure_path(id: @procedure.id)
-    end
   end
 
   def publish_validate
@@ -182,18 +145,6 @@ class Admin::ProceduresController < AdminController
     render layout: 'application'
   end
 
-  def monavis
-  end
-
-  def update_monavis
-    if !@procedure.update(procedure_params)
-      flash.now.alert = @procedure.errors.full_messages
-    else
-      flash.notice = 'le champ MonAvis a bien été mis à jour'
-    end
-    render 'monavis'
-  end
-
   def active_class
     @active_class = 'active'
   end
@@ -238,7 +189,7 @@ class Admin::ProceduresController < AdminController
   end
 
   def procedure_params
-    editable_params = [:libelle, :description, :organisation, :direction, :lien_site_web, :cadre_juridique, :deliberation, :notice, :web_hook_url, :euro_flag, :logo, :auto_archive_on, :monavis_embed]
+    editable_params = [:libelle, :description, :organisation, :direction, :lien_site_web, :cadre_juridique, :deliberation, :notice, :web_hook_url, :euro_flag, :logo, :auto_archive_on]
     permited_params = if @procedure&.locked?
       params.require(:procedure).permit(*editable_params)
     else
