--- app/views/admin/procedures/_informations.html.haml
@@ -1,142 +1,0 @@
-- if @procedure.locked?
-  .alert.alert-info
-    Cette démarche est publiée, certains éléments de la description ne sont plus modifiables
-
-.form-group
-  %h4 Libellé*
-  = f.text_field :libelle, class: 'form-control', placeholder: 'Libellé de la démarche'
-
-.form-group
-  %h4 Description*
-  = f.text_area :description, rows: '6', placeholder: 'Description du projet', class: 'form-control'
-
-- if !@procedure.locked?
-  .form-group
-    %h4 Conservation des données
-    = f.label :duree_conservation_dossiers_dans_ds, "Sur demarches-simplifiees.fr* (durée en mois après le début de l’instruction)"
-    = f.number_field :duree_conservation_dossiers_dans_ds, class: 'form-control', placeholder: '6', required: true
-
-    = f.label :duree_conservation_dossiers_hors_ds, "Hors demarches-simplifiees.fr* (durée en mois après la fin de l'instruction)"
-    = f.number_field :duree_conservation_dossiers_hors_ds, class: 'form-control', placeholder: '6', required: true
-
-- if @procedure.created_at.present?
-  .form-group
-    %h4 Où les usagers trouveront-ils le lien vers la démarche ?
-    = f.text_field :lien_site_web, class: 'form-control', placeholder: 'https://exemple.gouv.fr/ma_demarche'
-
-.form-group
-  %h4 Cadre juridique *
-  %p Texte qui justifie le droit de collecter les données demandées dans votre démarche auprès des usagers, par exemple :
-  %ul
-    %li Texte de loi (loi, décret, circulaire, arrêté,…)
-    %li Texte juridique (statuts, délibération, décision du conseil d'administration…)
-    %li
-      = link_to("En savoir plus avec cette vidéo de 5 minutes", CADRE_JURIDIQUE_URL, target: "_blank", rel: "noopener")
-
-  %p.help-block
-    %i.fa.fa-info-circle
-    Vous pouvez saisir un lien vers ce texte ou importer celui-ci directement.
-
-  .row
-    .col-md-6
-      = f.label :cadre_juridique, 'Lien vers le texte'
-      = f.text_field :cadre_juridique, class: 'form-control', placeholder: 'https://www.legifrance.gouv.fr/'
-    .col-md-6
-      = f.label :deliberation, 'Importer le texte'
-      - deliberation = @procedure.deliberation
-      - if !deliberation.attached?
-        = f.file_field :deliberation,
-          direct_upload: true
-      - else
-        %a{ href: url_for(deliberation), target: '_blank', rel: 'noopener' }
-          = deliberation.filename.to_s
-        - if @procedure.persisted?
-          = link_to 'supprimer', delete_deliberation_admin_procedure_path(@procedure), method: :delete
-        %br
-        Modifier :
-        = f.file_field :deliberation,
-          direct_upload: true
-
-.form-group
-  %h4 Notice explicative de la démarche
-  - notice = @procedure.notice
-  - if !notice.attached?
-    = f.file_field :notice,
-      direct_upload: true
-  - else
-    %a{ href: url_for(notice), target: '_blank', rel: 'noopener' }
-      = notice.filename.to_s
-    - if @procedure.persisted?
-      \-
-      = link_to 'supprimer', delete_notice_admin_procedure_path(@procedure), method: :delete
-    %br
-    Modifier :
-    = f.file_field :notice,
-      direct_upload: true
-
-.row
-  .col-md-6
-    %h4 Logo de la démarche
-    - if @procedure.logo.attached?
-      = image_tag @procedure.logo_url, { style: 'height: 40px; display: inline; margin-right: 6px;', id: 'preview_procedure_logo' }
-      \-
-
-      - if @procedure.persisted?
-        = link_to 'supprimer', delete_logo_admin_procedure_path(@procedure), method: :delete
-    = f.file_field :logo,
-      direct_upload: true,
-      accept: 'image/png, image/jpg, image/jpeg',
-      style: 'display: inline'
-
-    %div{ style: 'margin-top: 5px;' }
-      %i
-        Fichier accepté : JPG / JPEG / PNG
-  .col-md-6
-    %h4 Drapeau européen
-    .checkbox
-      %label
-        = f.check_box :euro_flag
-        Afficher le drapeau européen
-
-- if !@procedure.locked?
-  .row
-    .col-md-6
-      %h4 À qui s’adresse ma démarche ?
-      .checkbox
-        %label
-          = f.radio_button :for_individual, true
-          %b Ma démarche s’adresse à un particulier
-
-        %p
-          En choisissant cette option, l’usager devra renseigner son nom et prénom avant d’accéder au formulaire
-
-      .checkbox
-        %label
-          = f.radio_button :for_individual, false
-          %b Ma démarche s’adresse à une personne morale
-
-        %p
-          En choisissant cette option, l’usager devra renseigner son n° SIRET. Grâce à l’API Entreprise, seront alors automatiquement remontées les informations sur la personne morale type raison sociale ou adresse du siège social.
-
-      %b
-        Si votre démarche s’adresse indifféremment à une personne morale ou un particulier choisissez l'option "particuliers". Vous pourrez utilisez le champ SIRET directement dans le formulaire.
-
-.row
-  .col-md-6
-    %h4 Options avancées
-
-    - if feature_enabled?(:administrateur_web_hook)
-      %label{ for: :web_hook_url } Lien de rappel HTTP (webhook)
-      = f.text_field :web_hook_url, class: 'form-control', placeholder: 'https://callback.exemple.fr/'
-      %p.help-block
-        %i.fa.fa-info-circle
-        Vous pouvez définir un lien de rappel HTTP (aussi appelé webhook) pour notifier un service tiers du changement de l'état d’un dossier de cette démarche sur demarches-simplifiees.fr.
-        = link_to("Consulter la documentation du webhook", WEBHOOK_DOC_URL, target: "_blank", rel: "noopener")
-        \.
-
-    %label{ for: :auto_archive_on } Clôture automatique le
-    = f.date_field :auto_archive_on, id: 'auto_archive_on', value: @procedure.auto_archive_on
-    (à 00h01)
-    %p.help-block
-      %i.fa.fa-info-circle
-      La clôture automatique suspend la publication de la démarche et entraîne le passage de tous les dossiers au statut "en instruction", ce qui ne permet plus aux usagers de les modifier. Le passage en instruction des dossiers s'accompagne de l'envoi de l'email d'accusé de passage en instruction (configurable par l'administrateur dans la partie "E-mail" de la démarche).
