--- app/views/admin/procedures/edit.html.haml
@@ -1,6 +1,0 @@
-.row.white-back
-  #procedure_new.section.section-label
-    = form_for @procedure, url: url_for({ controller: 'admin/procedures', action: :update, id: @procedure.id }), multipart: true do |f|
-      = render partial: 'informations', locals: { f: f }
-      .text-right
-        = f.button 'Enregistrer', class: 'btn btn-success'
