--- app/views/admin/procedures/new.html.haml
@@ -1,9 +1,0 @@
-.row.white-back
-  %h2
-    = t('dynamics.admin.dossiers.tableau_de_bord.nouvelle_procedure')
-
-  #procedure_new.section.section-label
-    = form_for @procedure, url: { controller: 'admin/procedures', action: :create }, multipart: true do |f|
-      = render partial: 'informations', locals: { f: f }
-      .text-center
-        = f.button 'Valider', class: 'btn btn-info btn-lg btn-block', id: 'save-procedure'
