+++ app/views/new_administrateur/procedures/_monavis.html.haml
@@ -1,0 +1,15 @@
+%p.explication
+  Proposez aux usagers de donner un avis sur votre démarche. Pour ce faire, vous devez précédemment aller sur «
+  %a{ :href => "https://monavis.numerique.gouv.fr" } https://monavis.numerique.gouv.fr
+  », créer un compte, et référencer là démarche que vous venez de publier.
+  %br
+  %br
+  Vous pouvez
+  %a{ :href => "https://doc.demarches-simplifiees.fr/tutoriels/integration-du-bouton-mon-avis" } consulter notre tutoriel complet
+  pour intégrer le bouton « MonAvis » sur demarches-simplifiees.fr.
+  %br
+  %br
+  Une fois en possession du code généré sur le site MonAvis, vous pouvez le coller dans le champ ci-dessous :
+
+= f.label :monavis_embed, "Mon avis"
+= f.text_area :monavis_embed, rows: '6', placeholder: '<a href="https://monavis.numerique.gouv.fr/Demarches/123456?&view-mode=formulaire-avis&nd_mode=en-ligne-enti%C3%A8rement&nd_source=button&key=cd4a872d4"><img src="https://monavis.numerique.gouv.fr/monavis-static/bouton-bleu.png" alt="Je donne mon avis" title="Je donne mon avis sur cette démarche" /></a>', class: 'form-control'
